DCWL-2003-v2 moving bits and pieces
