+++ app/uk/gov/hmrc/customs/declarations/information/action/AuthAction.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.customs.declarations.information.action
 
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.errorBadRequest
-import uk.gov.hmrc.customs.declarations.information.controllers.CustomHeaderNames._
+import uk.gov.hmrc.customs.declarations.information.util.CustomHeaderNames._
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model.actionbuilders.{HasConversationId, HasRequest}
 import uk.gov.hmrc.customs.declarations.information.model.{AuthorisedAsCsp, BadgeIdentifier, Csp, Eori}
