+++ app/uk/gov/hmrc/customs/declarations/information/action/HeaderValidator.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.customs.declarations.information.action
 import play.api.mvc.Headers
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse._
-import uk.gov.hmrc.customs.declarations.information.controllers.CustomHeaderNames._
+import uk.gov.hmrc.customs.declarations.information.util.CustomHeaderNames._
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model.actionbuilders._
 import uk.gov.hmrc.customs.declarations.information.model.{BadgeIdentifier, ClientId, Eori}
