+++ app/uk/gov/hmrc/customs/declarations/information/action/InternalClientIdsCheckAction.scala
@@ -18,11 +18,11 @@ package uk.gov.hmrc.customs.declarations.information.action
 
 import play.api.mvc.{ActionRefiner, Result}
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.errorBadRequest
+import uk.gov.hmrc.customs.declarations.information.config.ConfigService
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model.DeclarationSubmissionChannel
 import uk.gov.hmrc.customs.declarations.information.model.actionbuilders.ActionBuilderModelHelper._
 import uk.gov.hmrc.customs.declarations.information.model.actionbuilders.{InternalClientIdsRequest, ValidatedHeadersRequest}
-import uk.gov.hmrc.customs.declarations.information.services.InformationConfigService
 
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
@@ -39,7 +39,7 @@ import scala.concurrent.{ExecutionContext, Future}
  */
 @Singleton
 class InternalClientIdsCheckAction @Inject()(val logger: InformationLogger,
-                                             val configService: InformationConfigService)(implicit ec: ExecutionContext)
+                                             val configService: ConfigService)(implicit ec: ExecutionContext)
   extends ActionRefiner[ValidatedHeadersRequest, InternalClientIdsRequest] {
   override def executionContext: ExecutionContext = ec
   val declarationSubmissionChannelErrorCode = "CDS60011"
