+++ app/uk/gov/hmrc/customs/declarations/information/controllers/DeclarationController.scala
@@ -25,6 +25,7 @@ import uk.gov.hmrc.customs.declarations.information.model._
 import uk.gov.hmrc.customs.declarations.information.model.actionbuilders.ActionBuilderModelHelper._
 import uk.gov.hmrc.customs.declarations.information.model.actionbuilders.AuthorisedRequest
 
+//TODO ???
 trait DeclarationController {
   def validateMrn(mrn: Mrn, logger: InformationLogger)(implicit asr: AuthorisedRequest[AnyContent]): Either[Result, Unit] = {
     if (mrn.validValue) {
