+++ test/unit/controllers/DeclarationFullControllerSpec.scala
@@ -29,6 +29,7 @@ import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse._
 import uk.gov.hmrc.customs.api.common.logging.CdsLogger
 import uk.gov.hmrc.customs.declarations.information.action.{ConversationIdAction, DeclarationFullAuthAction, DeclarationFullCheckAction, HeaderValidator, InternalClientIdsCheckAction, ShutterCheckAction, ValidateAndExtractHeadersAction}
+import uk.gov.hmrc.customs.declarations.information.config.{InformationConfig, ConfigService, InformationShutterConfig}
 import uk.gov.hmrc.customs.declarations.information.connectors.AbstractDeclarationConnector._
 import uk.gov.hmrc.customs.declarations.information.connectors.{ApiSubscriptionFieldsConnector, DeclarationFullConnector}
 import uk.gov.hmrc.customs.declarations.information.controllers.DeclarationFullController
@@ -52,7 +53,7 @@ import scala.concurrent.{ExecutionContext, Future}
 
 class DeclarationFullControllerSpec extends UnitSpec with Matchers with BeforeAndAfterEach {
   trait SetUp extends AuthConnectorStubbing {
-    protected val mockInformationConfigService: InformationConfigService = mock(classOf[InformationConfigService])
+    protected val mockInformationConfigService: ConfigService = mock(classOf[ConfigService])
     when(mockInformationConfigService.informationShutterConfig).thenReturn(InformationShutterConfig(Some(false), Some(false)))
     when(mockInformationConfigService.informationConfig).thenReturn(InformationConfig("url", 1, Seq()))
