+++ test/unit/services/declarations/DeclarationSearchServiceSpec.scala
@@ -24,6 +24,7 @@ import play.api.test.Helpers
 import play.mvc.Http.Status._
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.errorInternalServerError
+import uk.gov.hmrc.customs.declarations.information.config.{InformationConfig, ConfigService}
 import uk.gov.hmrc.customs.declarations.information.connectors.AbstractDeclarationConnector._
 import uk.gov.hmrc.customs.declarations.information.connectors.{ApiSubscriptionFieldsConnector, DeclarationSearchConnector}
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
@@ -54,7 +55,7 @@ class DeclarationSearchServiceSpec extends UnitSpec with BeforeAndAfterEach {
   protected lazy val mockDeclarationSearchConnector: DeclarationSearchConnector = mock(classOf[DeclarationSearchConnector])
   protected lazy val mockPayloadDecorator: BackendSearchPayloadCreator = mock(classOf[BackendSearchPayloadCreator])
   protected lazy val mockHttpResponse: HttpResponse = HttpResponse(OK, "<xml>some xml</xml>")
-  protected lazy val mockInformationConfigService: InformationConfigService = mock(classOf[InformationConfigService])
+  protected lazy val mockInformationConfigService: ConfigService = mock(classOf[ConfigService])
   protected lazy val mockInformationConfig: InformationConfig = mock(classOf[InformationConfig])
   protected val mrn = Mrn("theMrn")
   protected lazy val missingEoriResult = errorInternalServerError("Missing authenticated eori in service lookup").XmlResult.withConversationId
